backend/requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flask==3.0.2
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
werkzeug==3.0.1
zipp==3.18.1
Flask-Cors==4.0.0
flask-restx==1.3.0
PyJWT==2.6.0
SQLAlchemy==2.0.29
beautifulsoup4==4.12.3